+++ src/Products/urban/GenericLicence.py
@@ -84,7 +84,7 @@ optional_fields = [
     'roadMissingParts', 'roadMissingPartsDetails', 'locationMissingParts', 'locationMissingPartsDetails',
     'PRevU', 'prevuDetails', 'PRenU', 'prenuDetails', 'airportNoiseZone', 'airportNoiseZoneDetails',
     'description', 'rgbsr', 'rgbsrDetails', 'karstConstraints', 'karstConstraintsDetails',
-    'concentratedRunoffSRisk', 'concentratedRunoffSRiskDetails'
+    'concentratedRunoffSRisk', 'concentratedRunoffSRiskDetails', 'sevesoSite', 'natura2000'
 ]
 ##/code-section module-header
 
@@ -366,6 +366,35 @@ schema = Schema((
         schemata='urban_road',
         default_output_type='text/plain',
     ),
+<<<<<<< .courant
+=======
+    TextField(
+        name='sevesoSite',
+        allowable_content_types=('text/plain',),
+        widget=TextAreaWidget(
+            label='sevesosite',
+            label_msgid='urban_label_sevesoSite',
+            i18n_domain='urban',
+        ),
+        default_content_type='text/plain',
+        default_method='getDefaultText',
+        schemata='urban_road',
+        default_output_type='text/plain',
+    ),
+    LinesField(
+        name='natura2000',
+        widget=MultiSelectionWidget(
+            format='checkbox',
+            label='natura2000',
+            label_msgid='urban_label_natura2000',
+            i18n_domain='urban',
+        ),
+        schemata='urban_road',
+        multiValued=1,
+        vocabulary=UrbanVocabulary('natura_2000', inUrbanConfig=False),
+        default_method='getDefaultValue',
+    ),
+>>>>>>> .fusion-droit.r22982
     StringField(
         name='floodingLevel',
         widget=SelectionWidget(