+++ src/Products/urban/profiles/extra/config_default_values.py
@@ -1195,5 +1195,16 @@ default_values = {
             {'id': "medium", 'title': u"Moyen, bassin versant afférent entre 9 et 18 Ha"},
             {'id': "high", 'title': u"Elevé, bassin versant afférent de plus de 18 Ha"},
         ],
+
+        'seveso_site': [
+            'UrbanVocabularyTerm',
+            {'id': "exxonmobil", 'title': u"EXXONMOBIL PETROLEUM & CHEMICAL TOURNAI (Seuil bas)"},
+            {'id': "nyco", 'title': u"NYCO - STPC TOURNAI (Seuil bas)"},
+            {'id': "orica", 'title': u"ORICA - TITA NOBEL BELGIQUE S.A. TOURNAI (Seuil haut)"},
+            {'id': "univar", 'title': u"UNIVAR BELGIUM - QUARON TOURNAI (Seuil Haut)"},
+            {'id': "rynders", 'title': u"RYNDERS PETROLEUM BELGIUM ANTOING (seuil bas)"},
+            {'id': "transmat", 'title': u"TRANSMAT DEPOT ANTOING (seuil bas)"},
+        ],
+
     },
 }